+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.17.34-SNAPSHOT
 ### is.codion.common.core
 - ValueSetObserver added, related refactoring.
+- ValueSetObserver now extends Iterable.
 ### is.codion.framework.i18n
 - FrameworkMessages.confirmDeleteSelected() removed.
 ### is.codion.swing.framework.ui